+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [4-r.3] - 2021-06-10
+
+### Fixed
+
+* Fix motion event time value from Int to Float.
+
+
+## [4-r.3-beta.1] - 2021-05-13
+
+### Added
+
+* Implement a function to get the correct value when the time axis of the Bezier handle cannot be linear.
+
+
 ## [4-r.2] - 2021-03-09
 
 ### Fixed
@@ -43,5 +57,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Reformat code using Prettier and ESLint.
 
 
+[4-r.3]: https://github.com/Live2D/CubismWebFramework/compare/4-r.3-beta.1...4-r.3
+[4-r.3-beta.1]: https://github.com/Live2D/CubismWebFramework/compare/4-r.2...4-r.3-beta.1
 [4-r.2]: https://github.com/Live2D/CubismWebFramework/compare/4-r.1...4-r.2
 [4-r.1]: https://github.com/Live2D/CubismWebFramework/compare/ce2585a919ac6e99f64dd468933772c6f1abbcc7...4-r.1